Mostly minor updates to the create-CHM tutorial.
Changed the working directory setup per recent hyperspectral updates.

A couple of the links (lines 114, 128) had open parentheses between the first quote and {{ base.url; which I believe may have broken the links (the URL was all messed up on the website.

Added comment=NA to the final chunk so that the writeRaster() function is not commented out in the .R file.

A link at the bottom pointed to the wrong url, which was corrected here.